Add PR fixes 1
Fixes some issues that came up during review. The docker wrapper and the
Orchest resource manager has been put in the same module. The debug
functions use the same resource manager. Other minor fixes.
